requirements.txt
streamlit

# Utilities
numpy
pandas

# Face Recognition
scikit-learn
dlib-bin
git+https://github.com/ageitgey/face_recognition_models
setuptools<70.0.0

# Backend / Auth
supabase
bcrypt

# Others
segno
pillow